Use long for active connection count
Change GetActiveConnectionCount return type and local accumulator from int to long and remove the redundant cast when adding garnetServerBase.get_conn_active(). This prevents potential integer overflow when summing active connections across multiple server instances; callers may need to handle the updated long return value.

+++ libs/host/GarnetServer.cs
@@ -537,14 +537,14 @@ namespace Garnet
         /// <summary>
         /// Gets the current number of active connections directly from server instances.
         /// </summary>
-        private int GetActiveConnectionCount()
+        private long GetActiveConnectionCount()
         {
-            int count = 0;
+            long count = 0;
             if (servers != null)
             {
                 foreach (var garnetServerBase in servers.OfType<GarnetServerBase>())
                 {
-                    count += (int)garnetServerBase.get_conn_active();
+                    count += garnetServerBase.get_conn_active();
                 }
             }
             return count;
